Move Webpacker generation to app generator
